requirements.txt
absl-py==0.11.0
amqp==5.0.1
appnope==0.1.0
asgiref==3.2.7
astunparse==1.6.3
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
bleach==3.1.5
cached-property==1.5.2
cachetools==4.1.1
celery==5.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.7
django-crispy-forms==1.9.1
djangorestframework==3.12.1
entrypoints==0.3
form==0.0.1
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
jieba==0.42.1
Jinja2==2.11.2
joblib==0.15.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
kombu==5.0.2
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
nltk==3.5
notebook==6.0.3
numpy==1.18.4
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.0.4
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
redis==3.5.3
regex==2020.5.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
simple-pickle==1.5.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
stopwords==0.1.3
tabulate==0.8.7
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-process==2.5.2
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
urllib3==1.25.11
utils==1.0.1
vine==5.0.0
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.1.0
widgetsnbextension==3.5.1
wrapt==1.12.1
